+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- `esbuild-serve init` to generate an example index.html file.
 
 ## [0.3.4] - 2021-06-15
 Breaking: This version completely rewrites this library. Now it becomes a tiny